+++ src/librpbase/img/un-premultiply.cpp
@@ -1,6 +1,7 @@
 /***************************************************************************
  * ROM Properties Page shell extension. (librpbase)                        *
  * un-premultiply.cpp: Un-premultiply function.                            *
+ * Standard version. (C++ code only)                                       *
  *                                                                         *
  * Copyright (c) 2017-2018 by David Korth.                                 *
  *                                                                         *
@@ -32,10 +33,12 @@
 
 namespace LibRpBase {
 
-// Inverted pre-multiplication factors.
-// From Qt 5.9.1's qcolor.cpp.
-// These values are: 0x00FF00FF / alpha
-static const unsigned int qt_inv_premul_factor[256] = {
+/**
+ * Inverted pre-multiplication factors.
+ * From Qt 5.9.1's qcolor.cpp.
+ * These values are: 0x00FF00FF / alpha
+ */
+const unsigned int rp_image::qt_inv_premul_factor[256] = {
 	0, 16711935, 8355967, 5570645, 4177983, 3342387, 2785322, 2387419,
 	2088991, 1856881, 1671193, 1519266, 1392661, 1285533, 1193709, 1114129,
 	1044495, 983055, 928440, 879575, 835596, 795806, 759633, 726605,
@@ -71,34 +74,42 @@ static const unsigned int qt_inv_premul_factor[256] = {
 };
 
 /**
- * Un-premultiply an argb32_t pixel.
+ * Un-premultiply an argb32_t pixel. (Standard version)
+ * From qt-5.11.0's qrgb.h.
+ * qUnpremultiply()
+ *
  * This is needed in order to convert DXT2/3 to DXT4/5.
- * @param px	[in/out] argb32_t pixel to un-premultiply, in place.
+ *
+ * @param px	[in] ARGB32 pixel to un-premultiply.
+ * @return Un-premultiplied pixel.
  */
-static FORCEINLINE void un_premultiply_pixel(argb32_t &px)
+static FORCEINLINE uint32_t un_premultiply_pixel(uint32_t px)
 {
-	if (likely(px.a == 255)) {
-		// Do nothing.
-	} else if (px.a == 0) {
-		px.u32 = 0;
-	} else {
-		// Based on Qt 5.9.1's qUnpremultiply().
-		// (p*(0x00ff00ff/alpha)) >> 16 == (p*255)/alpha for all p and alpha <= 256.
-		const unsigned int invAlpha = qt_inv_premul_factor[px.a];
-		// We add 0x8000 to get even rounding.
-		// The rounding also ensures that qPremultiply(qUnpremultiply(p)) == p for all p.
-		px.r = (px.r * invAlpha + 0x8000) >> 16;
-		px.g = (px.g * invAlpha + 0x8000) >> 16;
-		px.b = (px.b * invAlpha + 0x8000) >> 16;
-	}
+	argb32_t rpx;
+	rpx.u32 = px;
+	if (likely(rpx.a == 255 || rpx.a == 0))
+		return px;
+
+	// Based on Qt 5.9.1's qUnpremultiply().
+	// (p*(0x00ff00ff/alpha)) >> 16 == (p*255)/alpha for all p and alpha <= 256.
+	const unsigned int invAlpha = rp_image::qt_inv_premul_factor[rpx.a];
+	// We add 0x8000 to get even rounding.
+	// The rounding also ensures that qPremultiply(qUnpremultiply(p)) == p for all p.
+	rpx.r = (rpx.r * invAlpha + 0x8000) >> 16;
+	rpx.g = (rpx.g * invAlpha + 0x8000) >> 16;
+	rpx.b = (rpx.b * invAlpha + 0x8000) >> 16;
+	return rpx.u32;
 }
 
 /**
  * Un-premultiply an ARGB32 rp_image.
+ * Standard version using regular C++ code.
+ *
  * Image must be ARGB32.
+ *
  * @return 0 on success; non-zero on error.
  */
-int rp_image::un_premultiply(void)
+int rp_image::un_premultiply_cpp(void)
 {
 	RP_D(const rp_image);
 	rp_image_backend *const backend = d->backend;
@@ -108,18 +119,90 @@ int rp_image::un_premultiply(void)
 		return -1;
 	}
 
-	// NOTE: SSE2 can't be used for un-premultiply due to lack of division instructions.
 	const int width = backend->width;
 	argb32_t *px_dest = static_cast<argb32_t*>(backend->data());
 	int dest_stride_adj = (backend->stride / sizeof(*px_dest)) - width;
 	for (int y = backend->height; y > 0; y--, px_dest += dest_stride_adj) {
 		int x = width;
 		for (; x > 1; x -= 2, px_dest += 2) {
-			un_premultiply_pixel(px_dest[0]);
-			un_premultiply_pixel(px_dest[1]);
+			px_dest[0].u32 = un_premultiply_pixel(px_dest[0].u32);
+			px_dest[1].u32 = un_premultiply_pixel(px_dest[1].u32);
 		}
 		if (x == 1) {
-			un_premultiply_pixel(*px_dest);
+			px_dest->u32 = un_premultiply_pixel(px_dest->u32);
+			px_dest++;
+		}
+	}
+	return 0;
+}
+
+/**
+ * Premultiply an argb32_t pixel. (Standard version)
+ * From qt-5.11.0's qrgb.h.
+ * qPremultiply()
+ *
+ * This is needed in order to use the Cairo graphics library.
+ *
+ * @param px	[in] ARGB32 pixel to premultiply.
+ * @return Premultiplied pixel.
+ */
+static FORCEINLINE uint32_t premultiply_pixel_inl(uint32_t px)
+{
+	const unsigned int a = (px >> 24);
+	if (likely(a == 255 || a == 0))
+		return px;
+
+	// Based on Qt 5.9.1's qPremultiply().
+	unsigned int t = (px & 0xff00ff) * a;
+	t = (t + ((t >> 8) & 0xff00ff) + 0x800080) >> 8;
+	t &= 0xff00ff;
+
+	px = ((px >> 8) & 0xff) * a;
+	px = (px + ((px >> 8) & 0xff) + 0x80);
+	px &= 0xff00;
+	return (px | t | (a << 24));
+}
+
+/**
+ * rp_image wrapper function for premultiply_pixel().
+ * @param px	[in] ARGB32 pixel to premultiply.
+ * @return Premultiplied pixel.
+ */
+uint32_t rp_image::premultiply_pixel(uint32_t px)
+{
+	return premultiply_pixel_inl(px);
+}
+
+/**
+ * Premultiply an ARGB32 rp_image.
+ *
+ * Image must be ARGB32.
+ *
+ * @return 0 on success; non-zero on error.
+ */
+int rp_image::premultiply(void)
+{
+	// TODO: Qt doesn't have SSE-optimized builds.
+
+	RP_D(const rp_image);
+	rp_image_backend *const backend = d->backend;
+	assert(backend->format == rp_image::FORMAT_ARGB32);
+	if (backend->format != rp_image::FORMAT_ARGB32) {
+		// Incorrect format...
+		return -1;
+	}
+
+	const int width = backend->width;
+	argb32_t *px_dest = static_cast<argb32_t*>(backend->data());
+	int dest_stride_adj = (backend->stride / sizeof(*px_dest)) - width;
+	for (int y = backend->height; y > 0; y--, px_dest += dest_stride_adj) {
+		int x = width;
+		for (; x > 1; x -= 2, px_dest += 2) {
+			px_dest[0].u32 = premultiply_pixel_inl(px_dest[0].u32);
+			px_dest[1].u32 = premultiply_pixel_inl(px_dest[1].u32);
+		}
+		if (x == 1) {
+			px_dest->u32 = premultiply_pixel_inl(px_dest->u32);
 			px_dest++;
 		}
 	}